Remove n_XHP_ENTITY from the XHPAST parse tree.
Summary: This node is redundant with n_XHP_TEXT and we currently emit confusing
and inconsistent parse trees for "<x>a &amp;</x>" vs "<x>&amp; a</x>". If you
care about the distinction the information is available at the token level so
there's no loss of generality in removing this node.

Test Plan:

Reviewers:

CC: