Fix RemarkupEngine issues with header markup, restore unit tests
Summary:
The header markup rule was pretty broken because of some changes I
made when porting this from trunk -- it would treat a very short header like
'=a=' as an h3, since we didn't break out of the loop, and fail entirely in
cases where we captured leading whitespace (this is an intentional change
to better accommodate certain parsers).

There were also some unit tests which I brought over but didn't actually hook
up. Hook those up and make the minor changes required to get them to pass.

Test Plan:
Ran the unit tests, looked at Maniphest content with header markup
in it.

Reviewed By: jungejason
Reviewers: jungejason, tuomaspelkonen
CC: jungejason, epriestley
Differential Revision: 90