Commit flex/bison scanner/parser to the repository
Summary:
Marcel suggested this as a good solution to the tricky install
process here: commit the generated scanner and parser to the repository so
users do not need to install flex and bison in order to build it. This is
explicitly permitted by the flex/bison licenses:

http://www.gnu.org/software/bison/manual/html_node/Conditions.html

Also detect and error/warn about bad versions of flex rather than exploding.

Test Plan:
Ran "make", "make parser scanner", "make cleanall", etc.

Reviewed By: aran
Reviewers: toulouse, jungejason, tuomaspelkonen, aran
CC: aran
Differential Revision: 444