Add some utf-8 building blocks to libphutil
Summary:
The performance of these is absolutely terrible but I think they work as
advertised, at least. The biggest problem with utf-8 is that PHP support for it
is terrible.

If we hit perf problems with this stuff I'll almost certainly need to bundle a
PHP extension with libphutil. This will open up some opportunities for us but
make HPHP integration and general configuration more difficult. :/

Test Plan:
Ran unit tests.

Reviewed By: jungejason
Reviewers: jungejason, aran, tuomaspelkonen
CC: aran, epriestley, jungejason
Differential Revision: 505