Added logging when 'pygmentize' fails.
Summary:
We are having some problems on 3/4 of our phabricator machines. The
pygmentize doesn't seem to work there for some reason. Let's log the
error and figure out what's going on.

Test Plan:
Renamed pygmentize binary on my sandbox and made sure the exception was logged.
Renamed it back and made sure a diff was syntax highlighted correctly.

Reviewed By: jungejason
Reviewers: jungejason, epriestley
Commenters: epriestley
CC: aran, jungejason, epriestley
Differential Revision: 663