Print query text from PhutilServiceProfile under the default echo handler (e.g., "arc --trace")
Summary: We print details about other kinds of service calls (conduit, exec),
but not queries. Show queries in "--trace" too. This doesn't matter too much for
"arc" since it doesn't use DB services, but it's helpful for "phd debug".

Test Plan: Ran "phd debug" with some daemon or other and got to see the queries
it was executing.

Reviewers: jungejason, nh, tuomaspelkonen, aran

Reviewed By: jungejason

CC: aran, jungejason

Differential Revision: 951