Ignore pygmentize errors when no lexer is found
Summary:
We get a lot of log spew from pygmentize not being able to find a lexer. This
reverts D663, which added logging for pygmentize.

Test Plan:
Loaded diffusion looking at a file with an extension for which there is no
lexer, and checked the error log to see that the error doesn't appear.

Reviewers: tuomaspelkonen, jungejason, epriestley

Reviewed By: epriestley

CC: aran, nh, epriestley, tuomaspelkonen

Differential Revision: 992