Don't show "OPTION SUMMARY" if only wildcard options exist in PhutilArgumentParser
Summary:
We don't print "wildcard" arugments in this section, but the test to remove them
is after we've already printed the header.

If you have a script which takes ONLY wildcard arguments, we print the header
and then an empty details section.

Test Plan: Viewed help of a command with only wildcard arguments.

Reviewers: davidreuss, btrahan

Reviewed By: davidreuss

CC: aran, epriestley

Maniphest Tasks: T831

Differential Revision: https://secure.phabricator.com/D1554