Build static xhpast
Summary:
Statically links the xhpast binary.  On my system, this makes it
increase in size from ~800kb to ~2000kb, which isn't a big deal.

This is also a horrible hack, but a quick one; we should really just be
more careful to build an xhpast that will actually run.

Test Plan: Built a working xhpast as described by the build script.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, epriestley

Maniphest Tasks: T953

Differential Revision: https://secure.phabricator.com/D1816