Fix monospaced-in-monospaced markup
Summary:
Using storeText() returns a token for the entire matched text which is later replaced back with the stored text.

So this means the "##" rule supercedes the "`" rule, in that:

  (1) ##`x`## -> <tt>`x`</tt>

...but:

  (2) `##x##` -> <tt><tt>x</tt></tt>

This seems OK though? Otherwise we could try to replace matches one by one from left to right I suppose.

Test Plan: See unit test.

Reviewers: vrana

Reviewed By: vrana

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D1896