Improve SSL error message for CURLE_COULDNT_RESOLVE_HOST
Summary:
This message doesn't narrow things down much, but provide a friendlier version of it.

Theoretically we could do an "are you connected to the internet?" test by getting a plurality vote on connectivity to several major web properties or something, but that seemed maybe-excessive?

Test Plan: Tried to connect to https://secure.phabricatoraodifnadfnlansldfnak.com/

Reviewers: tido, btrahan, vrana

Reviewed By: tido

CC: aran

Maniphest Tasks: T1105

Differential Revision: https://secure.phabricator.com/D2198