Don't let <span> cross lines when highlighting Remarkup
Summary:
If a layer above us wraps each line in some other element (like Paste or Differential) we'll end up with invalid markup if a highlighted element crosses a newline.

Instead, before we hit a newline, add a </span> and resume the span afterwards.

Test Plan: Created a paste with an @{article:...} link spanning a newline, got correct highlighting.

Reviewers: vrana, btrahan

Reviewed By: vrana

CC: aran

Differential Revision: https://secure.phabricator.com/D2261