Add HTTP queries to profiler
Summary:
My primary motivation is to see elasticsearch in DarkConsole.

I'm not sure if this is not too broad.
It will also make sense to add this to HTTPSFuture.
The consequence is that every Conduit call will have at least two entries in profiler. I'm not sure if it isn't too much noice.

Test Plan: Look at DarkConsole after searching with elasticsearch.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Koolvin

Differential Revision: https://secure.phabricator.com/D2464