Add PHP scrub for Pygmentize
Summary:
We don't add `<?php` in front of PHP code before sending it to Pygmentize.
The result is that PHP code is not highlighted if xhpast is not available or if there's an error in the code (in which case we fallback to Pygmentize).

Test Plan: Disabled xhpast, highlighted PHP.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D2813