Allow PhutilLocks to block
Summary: There are a couple of operations where I want to use locks now that we have a primitive, but when they're on web workflows we should block and wait for the lock. Allow locks to block.

Test Plan: Ran unit tests, "lock.php" with --wait, verified CPU usage for blocking processes is negligible, verified blocking processes acquire locks when they're released.

Reviewers: btrahan

Reviewed By: btrahan

CC: aran

Differential Revision: https://secure.phabricator.com/D2924