Allow caller to set a "preferred" editor for PhutilInteractiveEditor
Summary: The main use case for this is supporting an 'editor' configuration in arc. A fair number of users are unfamiliar with EDITOR or with environmental variables in general, and setting them on Windows is a particular pain. Instead, allow arc to have an editor config so we can simplify instructions for "how to set up editors".

Test Plan: Added arc support for this, called `arc set-config editor 'mate -w'`, ran 'arc diff', got textmate instead of nano.

Reviewers: btrahan

Reviewed By: btrahan

CC: aran

Maniphest Tasks: T1309

Differential Revision: https://secure.phabricator.com/D2955