Make error message from STDERR in exec channel
Summary:
If client prints to stderr then we fail horribly.
Our tests print to stderr quite often and there's not much I can do about it.
One example is when we throw in destructor (which is not supported by HPHP).

This is not ideal because message comes out of order.
I'll probably reorder STDERR before any input.

Test Plan:
  fprintf(STDERR, 'Test.'); // From client.

Reviewers: vrana

Reviewed By: vrana

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D3213