Allow intra-domain links in Remarkup
Summary:
When parsing a link of the form `[[/path/...]]`, the protocol
gets marked as `''`, which is not allowed, so the link doesn't get
linkified and later turns into a Phriction link. The previous line
specifically allows this case, so it's a bit silly. So if the URI starts
with `/`, skip the protocol check.

Test Plan: Try linking things in Differential comments.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, Korvin, starruler

Differential Revision: https://secure.phabricator.com/D3281