Always use SSLv3 for HTTPSFuture
Summary:
A couple of users are encoutering an issue which is fixed by adding an explicit SSLv3:

https://github.com/facebook/libphutil/issues/17

Since v3 is 16 years old and I can't find anyone on the internet saying that swithching to v2 solved a problem, let's see if we can get away without adding a bunch of configuration for this or writing our own SSL library.

Test Plan: Made this diff. Googled for discussion of this option. :/

Reviewers: vrana, btrahan

Reviewed By: vrana

CC: aran

Differential Revision: https://secure.phabricator.com/D3767