Fix an issue where the first service call on the page may not be marked complete if it is a kvcache request
Summary: The first call ID we issue is 0. Possibly we should start at 1, but the ID counter is slightly complicated to muck with. Do a strict null test instead.

Test Plan: Loaded a page which did a kvcache hit first, didn't see a darkconsole issue with undefined 'duration' key in service log.

Reviewers: btrahan, codeblock

Reviewed By: codeblock

CC: aran

Differential Revision: https://secure.phabricator.com/D4358