use parse_url() for URI parsing
Summary:
We periodically have an issue with really long URLs hitting a
pcre backtracking limit (hhvm has a low limit configured due to historical
reasons).  Avoiding PCRE in the URL parser seems like a good idea.

Test Plan:
`arc unit` passes, but with the caveat that I've only run this
on hhvm so far; will arc patch it down to a mac shortly.  The concern is
whether parse_url groks svn+ssh and this may be PHP version dependent.

Reviewers: nh, epriestley

Reviewed By: epriestley

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D5726