Allow named links inside simple tables in Remarkup
Summary:
Fixes T3064. Currently, the simple table syntax sees the `|` in the middle of a named link as a cell divider.

Special-case this since I don't have a clean way of fixing it otherwise, but it seems reasonable to fix it.

Test Plan:
Added a passing test.

{F41831}

Reviewers: vrana, btrahan

Reviewed By: btrahan

CC: aran

Maniphest Tasks: T3064

Differential Revision: https://secure.phabricator.com/D5798