Provide the ability to set an alter-operation cost to PhutilEditDistanceMatrix
Summary:
In Arcanist, we currently price altering the edit operation so that strings like `(x)` vs `((x))` get analyzed as `isssi` instead of `issis` (that is, the first and last parens are highlighted as "new", rather than the first open and first close paren) which is better for human viewers.

Support this in PhutilEditDistanceMatrix.

Test Plan: Provided an alter cost in `arc` and ran the tests. They pass after this change.

Reviewers: btrahan, Afaque_Hussain

Reviewed By: btrahan

CC: aran

Maniphest Tasks: T2379

Differential Revision: https://secure.phabricator.com/D6013