Encode/sign JIRA POST requests like JIRA expects
Summary: Ref T3687. Currently we use generic OAuth1 rules here, but JIRA expects bodies to be JSON encoded and not included as part of the signature. Left a comment about this in case we need some other behavior in general, but this is probably the simplest thing that could possibly work for now.

Test Plan: Performed API POST requests to JIRA, got action on the other side.

Reviewers: btrahan

Reviewed By: btrahan

CC: aran

Maniphest Tasks: T3687

Differential Revision: https://secure.phabricator.com/D6891